chore: Extract VssSignal + VssBranch interface
